LPS-60393 Rename to dynamic-data-mapping-form-field-type
